Introduce playRound function with two param
	The playRound function includes two parameters, playSelection and
computerSelection. The function has if-else conditions, which compares the
choice from player and computer (both being lowerCase). After comparing, a string message is declared and stored in a result variable, which displays if the player won or lost and by what choice. Then the result value is returned in the function.