Update JS file for choice and score incrementation
	Added multiple elements of HTML to JS file as constants using Query
Selector. Const variables for choices (RPS), 2 score display divs and play again button. Through the use of addEvenListener, when player clicks on any RPS div
choices - the score is incremented based on the winner. All three RPS choices divs have a addEvenlistener with individual Protocol function. The function gets
computer choice and compares it with the player choice to provide the result.